+++ RELEASES.md
@@ -734,3 +734,4 @@
 -----------------
   - refactor: prioritization of data providers by length of filter pattern desc
   - fix: when using volatile permissions, system permissions starting with _ are retained in same datastore as groups
+  - fix: client connection status is set to ACTIVE before authentication happens, added new status RECONNECT_ACTIVE